+++ sktime_dl/classification/_cnn.py
@@ -5,7 +5,7 @@
 __author__ = "James Large"
 
 from sktime_dl.deeplearning.base.estimators import BaseDeepClassifier
-from networks._cnn import CNNNetwork
+from sktime_dl.networks._cnn import CNNNetwork
 from sktime_dl.utils import check_and_clean_data, \
     check_and_clean_validation_data
 from sklearn.utils import check_random_state
