+++ sktime_dl/networks/__init__.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+__all__ = [
+    "BaseDeepRegressor",
+    "CNNRegressor",
+]
+
+from sktime_dl.networks._network import BaseDeepNetwork
+from sktime_dl.networks._cnn import CNNNetwork
