+++ sktime_dl/regression/__init__.py
@@ -1,6 +1,8 @@
 # -*- coding: utf-8 -*-
 __all__ = [
+    "BaseDeepRegressor",
     "CNNRegressor",
 ]
 
+from sktime_dl.regression._regressor import BaseDeepRegressor
 from sktime_dl.regression._cnn import CNNRegressor
